Add browse movies fragment
